requirements.txt
django==3.2.2
djangorestframework==3.12.2
django-cors-headers==3.7.0
daphne==3.0.1
drf-yasg2==1.19.4
celery==5.0.5
beautifulsoup4==4.9.3
pyowm==3.2.0
psycopg2==2.8.6
openpyxl==3.0.5